Refactored searchProducts
