+++ src/components/profileaddresses.main.jsx
@@ -24,11 +24,8 @@ import ReactRouterPropTypes from 'react-router-prop-types';
 import intl from 'react-intl-universal';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router';
-import { login } from '../utils/AuthService';
-import cortexFetch from '../utils/Cortex';
+import { deleteUri } from '../utils/AuthService';
 import './profileaddresses.main.less';
-
-const Config = require('Config');
 
 class ProfileAddressesMain extends React.Component {
   static propTypes = {
@@ -48,21 +45,14 @@ class ProfileAddressesMain extends React.Component {
   }
 
   handleDelete(link) {
-    login().then(() => {
-      cortexFetch(link, {
-        method: 'delete',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: localStorage.getItem(`${Config.cortexApi.scope}_oAuthToken`),
-        },
-      }).then(() => {
+    deleteUri(link)
+      .then(() => {
         const { onChange } = this.props;
         onChange();
       }).catch((error) => {
         // eslint-disable-next-line no-console
         console.error(error.message);
       });
-    });
   }
 
   renderAddresses() {
